move assests and hooks in util folder
