+++ packages/ui-toolkit/src/components/atoms/Toaster/utils/assets.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { ToastTypes } from './types';
+import { ToastTypes } from '../types';
 
 export const getAsset = (type: ToastTypes): JSX.Element | null => {
   switch (type) {
